Simplify scraper to run iterative for smaller batches

+++ r/main_marktplaats_scraper.R
@@ -63,7 +63,7 @@ if(FALSE) {
       number_of_tries = 3, # in case of connection time-outs
       scrape_interval = 4, # interval between two scrapes, in hours
       # BigQuery settings
-      batch_size = 100
+      batch_size = 500
     )
   )
 }
@@ -88,59 +88,41 @@ open_ads <- get_ads_from_bigquery(
 tryCatch(expr = {
   listed_ads <- list_advertisements(
     url = settings$search_url,
-    advertisement_type = "individuals" #,max_pages = 3
+    advertisement_type = "individuals"#,max_pages = 3
   )
 },finally = {
-  listed_ads <- data.frame(ad_id = open_ads$ad_id[1])
+  if(!exists("listed_ads")) {
+    listed_ads <- data.frame(ad_id = "m1234567890")
+  }
 })
 
-# Determine which ads to scrape
-ads_to_scrape <- determine_ads_to_scrape(
-  ads_listed = listed_ads, 
-  ads_seen = open_ads,
-  scrape_interval_h = settings$scrape_interval
-)
-
-# Create batches
-batches <- split(ads_to_scrape, ceiling(seq_along(ads_to_scrape)/settings$batch_size))
-# Create empty results table
-scraped_ads <- data.frame()
-
-# Scrape and upload them per batch
-for(batch in batches) {
-  try({
-    # Get all open ads from google bigquery
-    to_scrape <- get_ads_from_bigquery(
-      project = settings$project,
-      bq_dataset = settings$bq_dataset,
-      bq_table = settings$bq_table,
-      method = "open"
+# Determine which ads to scrape and scrape them!
+tryCatch(expr = {
+  scraped_ads <- determine_ads_to_scrape(
+      ads_listed = listed_ads, 
+      ads_seen = open_ads,
+      scrape_interval_h = settings$scrape_interval,
+      max_n = settings$batch_size
     ) %>% 
-      dplyr::filter(ad_id %in% batch) %>% 
-      dplyr::mutate(time_since_last_scrape = difftime(Sys.time(),last_scrape,units="hours")) %>% 
-      dplyr::filter(time_since_last_scrape >= settings$scrape_interval) %>% 
-      dplyr::arrange(-time_since_last_scrape)
-    # Scrape batch of ads
-    if(dim(to_scrape)[1] > 0) {
-      scraped <- scrape_ads(
-        ad_ids = to_scrape$ad_id,
-        ads_per_minute = settings$ads_per_minute,
-        report_every_nth_scrape = settings$report_every_nth_scrape,
-        number_of_tries = settings$number_of_tries
-      )
-      # Upload batch of scraped ads to bigquery
-      upload_ads_to_bigquery(
-        scraped_ads = scraped,
-        project = settings$project,
-        bq_dataset = settings$bq_dataset,
-        bq_table = settings$bq_table,
-        batch_size = settings$batch_size
-      )
-      # Add them to total set
-      scraped_ads <- dplyr::bind_rows(scraped_ads,scraped)
-    }
-  })
-}
+    scrape_ads(
+      ads_per_minute = settings$ads_per_minute,
+      report_every_nth_scrape = settings$report_every_nth_scrape,
+      number_of_tries = settings$number_of_tries
+    )
+},finally = {
+  if(!exists("scraped_ads")) {
+    scraped_ads <- data.frame()
+  }
+})
+
+# Upload scraped ads to bigquery
+upload_ads_to_bigquery(
+  scraped_ads = scraped_ads,
+  project = settings$project,
+  bq_dataset = settings$bq_dataset,
+  bq_table = settings$bq_table,
+  batch_size = settings$batch_size
+)
 
 # Add log items
 log_items$n_rows_scraped <- nrow(scraped_ads)
@@ -150,53 +132,58 @@ log_items$n_existing_ads <- sum(scraped_ads$ad_id %in% open_ads$ad_id)
 log_items$end_time_scraping <- Sys.time()
 
 
-#### Determine ad images to scrape ####
-# Set scope
-options(googleAuthR.scopes.selected = c(settings$scope))
-# Authentication
-googleCloudStorageR::gcs_auth()
-# Get bucket and bucket info
-bucket <- googleCloudStorageR::gcs_list_buckets(settings$project)
-bucket_info <- googleCloudStorageR::gcs_get_bucket(bucket$name)
-# Get bucket objects
-#objects <- googleCloudStorageR::gcs_list_objects(bucket$name) # NOT CORRECTLY IMPLEMENTED, LIMITED TO 1000 RESULTS!
-objects <- get_add_images_from_gcloud(
-  bucket_name = bucket$name,
-  prefix = settings$imageDir
-)
+#### Scrape and upload ad images ####
+# Create dir for temp storage of images
+dir.create(settings$imageDir)
+# Run within a tryCatch
+tryCatch(expr = {
+  # Set scope
+  options(googleAuthR.scopes.selected = c(settings$scope))
+  # Authentication
+  googleCloudStorageR::gcs_auth()
+  # Get bucket and bucket info
+  bucket <- googleCloudStorageR::gcs_list_buckets(settings$project)
+  bucket_info <- googleCloudStorageR::gcs_get_bucket(bucket$name)
+  # Get bucket objects
+  #objects <- googleCloudStorageR::gcs_list_objects(bucket$name) # NOT CORRECTLY IMPLEMENTED, LIMITED TO 1000 RESULTS!
+  objects <- get_add_images_from_gcloud(
+    bucket_name = bucket$name,
+    prefix = settings$imageDir
+  )
+  
+  # Create filter for images which have already been collected
+  gathered_images <- data.frame(name = objects) %>%
+    dplyr::filter(grepl(settings$imageDir,name)) %>% 
+    dplyr::mutate(ad_id = as.character(stringr::str_extract(name,'[am][0-9]{1,10}'))) %>% 
+    dplyr::filter(grepl("[mM|aA][0-9]{9}", ad_id)) %>% 
+    dplyr::select(ad_id) %>% 
+    dplyr::distinct(.keep_all = T)
+  
+  # Get ads to collect images for
+  ads_with_images <- get_ads_with_images(
+    project = settings$project,
+    bq_dataset = settings$bq_dataset,
+    bq_table = settings$bq_table
+  )
+  # Apply filter
+  ads_to_get_images_for <- ads_with_images[!ads_with_images %in% gathered_images$ad_id]
 
-# Create filter for images which have already been collected
-gathered_images <- data.frame(name = objects) %>%
-  dplyr::filter(grepl(settings$imageDir,name)) %>% 
-  dplyr::mutate(ad_id = as.character(stringr::str_extract(name,'[am][0-9]{1,10}'))) %>% 
-  dplyr::filter(grepl("[mM|aA][0-9]{9}", ad_id)) %>% 
-  dplyr::select(ad_id) %>% 
-  dplyr::distinct(.keep_all = T)
-
-# Get ads to collect images for
-ads_with_images <- get_ads_with_images(
-  project = settings$project,
-  bq_dataset = settings$bq_dataset,
-  bq_table = settings$bq_table
-)
-# Apply filter
-ads_to_get_images_for <- ads_with_images[!ads_with_images %in% gathered_images$ad_id]
+  # Scrape and upload all images
+  list(
+    ad_id = ads_to_get_images_for,
+    file_path = settings$imageDir,
+    bucket_name = bucket$name
+  ) %>% 
+    purrr::pwalk(upload_ad_images_to_gcloud)
+  
+},finally = {
+  if(!exists("ads_to_get_images_for")) {
+    ads_to_get_images_for <- character()
+  }
+})
 
 # Add log items
 log_items$n_new_ads_with_images <- length(ads_to_get_images_for)
-
-
-#### Scrape and upload ad images ####
-dir.create(settings$imageDir)
-# Scrape and upload all images
-list(
-  ad_id = ads_to_get_images_for,
-  file_path = settings$imageDir,
-  bucket_name = bucket$name
-) %>% 
-  purrr::pwalk(upload_ad_images_to_gcloud)
-
-# Add log items
 log_items$n_new_images <- length(list.files(settings$imageDir))
 log_items$end_time_uploading_images <- Sys.time()
 duration_in_mins <- function(start,end) paste0(round(as.numeric(difftime(start,end,units="mins")),1),' minutes')
@@ -205,7 +192,9 @@ log_items$duration_uploading_images <- duration_in_mins(log_items$end_time_uploa
 log_items$total_time <- duration_in_mins(log_items$end_time_uploading_images,log_items$start_time)
 
 # Cleanup
-unlink(settings$imageDir, recursive=TRUE)
+if(file.exists(settings$imageDir)) {
+  unlink(settings$imageDir, recursive=TRUE)
+}
 
 
 #### Upload logs ####
